requirements.txt
# Core dependencies
pdfplumber
python-docx
faiss-cpu
numpy
sentence-transformers
torch
open-clip-torch
transformers
pytesseract
whisper
llama-cpp-python
Pillow

# API & Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Production
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Logging & Monitoring
python-json-logger>=2.0.7

# Async
anyio>=4.0.0

